.agent/skills/youtube-scraping/scripts/requirements.txt
yt-dlp
youtube-transcript-api
lingua-language-detector
